updated stop script

+++ scripts/stop_server.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
 SIMPLE_REDDIT_PID=$(ps -A -o pid,cmd|grep simple-reddit-build | grep -v grep |head -n 1 | awk '{print $1}')
-kill -9 $SIMPLE_REDDIT_PID
+echo $SIMPLE_REDDIT_PID
+if [ -n $SIMPLE_REDDIT_PID ]; then
+    kill -9 $SIMPLE_REDDIT_PID
+fi
